cismatch_backend/requirements.txt
Django==5.1.2
psycopg2-binary==2.9.10
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
channels==4.0.0
python-a2s==1.4.0
social-auth-app-django==5.4.2
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.1
yookassa==3.5.0
steam==1.4.0
steam-client
eventemitter==0.2.0
gevent==24.11.1
protobuf==3.20.*
steampy
uvicorn==0.20.0
daphne==4.0.0
asgiref==3.8.1
fastapi==0.115.11